pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.0",
    "fastapi[standard]>=0.115.3",
    "motor>=3.6.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.9.0",
    "ruff>=0.7.1",
    "structlog>=24.4.0",
    "uvicorn>=0.32.0",
]
